Created the registerWithSchool page
A simple page where students from RPI can log in using CAS and non-RPI students can login elsewhere. As we begin to integrate these pages with the backend, there are some things that we need to be thoughtful about. For instance, we need to make sure CAS works for RPI students and that we have individual login pages for any other schools/groups that are registered with Poll Buddy.